kmeans/requirements.txt
Pillow==1.1.7
Matplotlib==3.0.3
